+++ c/obd_dtc_row_update_0x64258.c
@@ -25,7 +25,7 @@
 void obd_dtc_row_update_0x64258(void)
 {
     uint16_t row = *(volatile uint16_t *)0xFFFF8D74;
-    uint8_t *p = (uint8_t *)(0xFFFF8930 + (uint32_t)row * 0x34);
+    uint8_t *p = (uint8_t *)(uintptr_t)(0xFFFF8930 + (uint32_t)row * 0x34);
     p[0x32] = (uint8_t)(p[0x32] + p[0x07] + 0xFF);
     p[0x07] = 1;
     p[0x32] = (uint8_t)(p[0x32] + p[0x08] + 0xF9);
